+++ CHANGELOG.md
@@ -46,6 +46,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 興行スケジュールのsuperEventを最適化
 - 経理レポートを最適化
 - 注文を最適化
+- SendActionのIRecipientを最適化
 
 ### Deprecated